fix: 피드백 반영 (#39) - API 수정 - DB 스키마 및 DML 수정 - 비관적 락 수정 - 메소드 명 수정

+++ src/main/java/com/flab/tabling/waiting/service/WaitingService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import com.flab.tabling.global.exception.ErrorCode;
 import com.flab.tabling.member.domain.Member;
 import com.flab.tabling.store.domain.Store;
-import com.flab.tabling.waiting.domain.Status;
 import com.flab.tabling.waiting.domain.Waiting;
+import com.flab.tabling.waiting.domain.WaitingStatus;
 import com.flab.tabling.waiting.exception.WaitingDuplicatedException;
 import com.flab.tabling.waiting.exception.WaitingExceededException;
 import com.flab.tabling.waiting.exception.WaitingNotFoundException;
@@ -22,7 +22,7 @@ public class WaitingService {
 
 	public Waiting add(Store store, Member member, Integer headCount) {
 		checkWaitingQueueFull(store);
-		if (waitingRepository.findByMemberAndStoreAndStatus(member, store, Status.WAITING).isPresent()) {
+		if (waitingRepository.findByMemberAndStoreAndStatus(member, store, WaitingStatus.ONGOING).isPresent()) {
 			throw new WaitingDuplicatedException(ErrorCode.PARAMETER_DUPLICATED,
 				"waiting for given store already exists");
 		}
@@ -30,7 +30,7 @@ public class WaitingService {
 			.store(store)
 			.member(member)
 			.headCount(headCount)
-			.status(Status.WAITING)
+			.waitingStatus(WaitingStatus.ONGOING)
 			.build();
 		try {
 			waitingRepository.save(waiting);
@@ -40,30 +40,33 @@ public class WaitingService {
 		}
 		return waiting;
 	}
-	
-	public void cancelMember(Store store, Member member) {
-		Waiting waiting = waitingRepository.findByMemberAndStoreAndStatus(member, store, Status.WAITING)
+
+	public void cancelMember(Store store, Member member, Long waitingId) {
+		Waiting waiting = waitingRepository.findByMemberAndStoreAndStatus(member, store, WaitingStatus.ONGOING)
 			.orElseThrow(() -> new WaitingNotFoundException(ErrorCode.PARAMETER_NOT_FOUND,
 				"waiting for given store not found"));
+		if (!waitingId.equals(waiting.getId())) {
+			throw new WaitingNotFoundException(ErrorCode.PARAMETER_NOT_FOUND, "The given waiting is not yours");
+		}
 		waitingRepository.delete(waiting);
 	}
 
 	public void cancelFirst(Store store) {
-		Waiting waiting = waitingRepository.findFirstByStoreAndStatus(store, Status.WAITING)
+		Waiting waiting = waitingRepository.findFirstByStoreAndStatusOrderByCreatedAt(store, WaitingStatus.ONGOING)
 			.orElseThrow(() -> new WaitingNotFoundException(ErrorCode.STORE_NOT_FOUND,
 				"waiting for given store is empty"));
 		waitingRepository.delete(waiting);
 	}
 
 	public void acceptFirst(Store store) {
-		Waiting waiting = waitingRepository.findFirstByStoreAndStatusByPessimisticLock(store, Status.WAITING)
+		Waiting waiting = waitingRepository.findFirstByStoreAndStatus(store, WaitingStatus.ONGOING)
 			.orElseThrow(() -> new WaitingNotFoundException(ErrorCode.STORE_NOT_FOUND,
 				"waiting for given store is empty"));
 		waiting.accept();
 	}
 
 	private void checkWaitingQueueFull(Store store) {
-		Integer count = waitingRepository.countWithPessimisticLockByStoreAndStatus(store, Status.WAITING);
+		Integer count = waitingRepository.countWithPessimisticLockByStoreAndStatus(store, WaitingStatus.ONGOING);
 		if (count >= store.getMaxWaitingCount()) {
 			throw new WaitingExceededException(ErrorCode.INVALID_PARAMETER, "waiting queue is full");
 		}
